+++ src/test/java/net/kaczmarzyk/spring/data/jpa/Customer.java
@@ -45,6 +45,9 @@ public class Customer {
     @Enumerated(EnumType.STRING)
     private Gender genderAsString;
 
+    private char genderAsChar;
+    private Character genderAsCharacter;
+
     private String firstName;
 
     private String lastName;
@@ -95,6 +98,13 @@ public class Customer {
     @OneToMany(mappedBy = "customer", fetch = FetchType.LAZY, cascade = {CascadeType.PERSIST, CascadeType.REMOVE})
     private Set<Badge> badges;
 
+    @ElementCollection
+    @CollectionTable(name = "customer_phone_numbers", joinColumns = @JoinColumn(name = "id"))
+    private Set<String> phoneNumbers;
+
+    @ElementCollection
+    @CollectionTable(name = "customer_lucky_numbers", joinColumns = @JoinColumn(name = "id"))
+    private Set<Long> luckyNumbers;
 
     public Customer() {
     }
@@ -103,6 +113,8 @@ public class Customer {
         this.firstName = firstName;
         this.lastName = lastName;
         this.gender = gender;
+        this.genderAsChar = gender.name().charAt(0);
+        this.genderAsCharacter = this.genderAsChar;
         this.genderAsString = gender;
         this.registrationDate = registrationDate;
         this.address.setStreet(street);
@@ -131,6 +143,8 @@ public class Customer {
     public void setGender(Gender gender) {
         this.gender = gender;
         this.genderAsString = gender;
+        this.genderAsChar = gender.name().charAt(0);
+        this.genderAsCharacter = this.genderAsChar;
     }
 
     public Timestamp getLastSeen() {
@@ -288,4 +302,25 @@ public class Customer {
         return badges;
     }
 
+    public Set<String> getPhoneNumbers() {
+        if (phoneNumbers == null) {
+            phoneNumbers = new HashSet<>();
+        }
+        return phoneNumbers;
+    }
+
+    public Set<Long> getLuckyNumbers() {
+        if (luckyNumbers == null) {
+            luckyNumbers = new HashSet<>();
+        }
+        return luckyNumbers;
+    }
+
+    public  void addLuckyNumber(Long luckyNumber) {
+        getLuckyNumbers().add(luckyNumber);
+    }
+
+    public void addPhoneNumber(String phoneNumber) {
+        getPhoneNumbers().add(phoneNumber);
+    }
 }